+++ src/lib/data/skills.ts
@@ -3,19 +3,6 @@ export type Skill = {
   color: string;
 };
 
-export const getSkillsByName = (
-  names: string[] | undefined
-): Skill[] | undefined => {
-  if (!names) {
-    return undefined;
-  }
-  return names
-    .map((name) =>
-      allSkills.find((skill) => skill.name.toLowerCase() === name.toLowerCase())
-    )
-    .filter((skill): skill is Skill => skill !== undefined);
-};
-
 export const skills: Skill[] = [
   { name: "Java", color: "#df8e2b" },
   { name: "JavaScript", color: "#f0db4f" },
