+++ src/lib/queries/skills.ts
@@ -1,0 +1,14 @@
+import { allSkills, Skill } from "../data/skills";
+
+export const getSkillsByName = (
+  names: string[] | undefined
+): Skill[] | undefined => {
+  if (!names) {
+    return undefined;
+  }
+  return names
+    .map((name) =>
+      allSkills.find((skill) => skill.name.toLowerCase() === name.toLowerCase())
+    )
+    .filter((skill): skill is Skill => skill !== undefined);
+};
